controllers change, voter

+++ src/Controller/ChildrenController.php
@@ -5,26 +5,16 @@ namespace App\Controller;
 use App\Entity\Child;
 use App\Form\ChildType;
 use App\Repository\ChildRepository;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 /**
- * @Route("/children")
+ * @Route("/profil/children")
  */
 class ChildrenController extends AbstractController
 {
-    /**
-     * @Route("/", name="app_children_index", methods={"GET"})
-     */
-    public function index(ChildRepository $childRepository): Response
-    {
-        return $this->render('children/index.html.twig', [
-            'children' => $childRepository->findAll(),
-        ]);
-    }
-
     /**
      * @Route("/new", name="app_children_new", methods={"GET", "POST"})
      */
@@ -35,9 +25,10 @@ class ChildrenController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $child->setUser($this->getUser());
             $childRepository->add($child, true);
 
-            return $this->redirectToRoute('app_children_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('app_user', [], Response::HTTP_SEE_OTHER);
         }
 
         return $this->renderForm('children/new.html.twig', [
@@ -51,6 +42,9 @@ class ChildrenController extends AbstractController
      */
     public function show(Child $child): Response
     {
+        // Voter Control
+        $this->denyAccessUnlessGranted('CHILD_ACCESS', $child);
+
         return $this->render('children/show.html.twig', [
             'child' => $child,
         ]);
@@ -61,13 +55,16 @@ class ChildrenController extends AbstractController
      */
     public function edit(Request $request, Child $child, ChildRepository $childRepository): Response
     {
+        // Voter Control
+        $this->denyAccessUnlessGranted('CHILD_ACCESS', $child);
+
         $form = $this->createForm(ChildType::class, $child);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $childRepository->add($child, true);
 
-            return $this->redirectToRoute('app_children_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('app_user', [], Response::HTTP_SEE_OTHER);
         }
 
         return $this->renderForm('children/edit.html.twig', [
@@ -81,10 +78,13 @@ class ChildrenController extends AbstractController
      */
     public function delete(Request $request, Child $child, ChildRepository $childRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$child->getId(), $request->request->get('_token'))) {
+        // Voter Control
+        $this->denyAccessUnlessGranted('CHILD_ACCESS', $child);
+
+        if ($this->isCsrfTokenValid('delete' . $child->getId(), $request->request->get('_token'))) {
             $childRepository->remove($child, true);
         }
 
-        return $this->redirectToRoute('app_children_index', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('app_user', [], Response::HTTP_SEE_OTHER);
     }
 }
